==> 1_Initiator.ipynb <==
import numpy as np
import pandas as pd
import spacy
from spacy import displacy
import sys

# spaCy for POS tagging
nlp=spacy.load("fr")


lexicalclues=pd.read_excel(
    "./excels/0_Phase3_LexicalCluesENFR.xlsx")

lexicalclues["Rule"]=lexicalclues["Direction"]+"|"+lexicalclues["Trigger Type"]+"|"+lexicalclues["Modifier"]+"|"+lexicalclues["Window Size"].astype(str)

lexicalclues_with_x=lexicalclues[~(lexicalclues["Grouping"].str.contains("X").fillna(False))]
lexicalclues_without_x=lexicalclues[(lexicalclues["Grouping"].str.contains("X").fillna(False))]

lexicalclues=lexicalclues_with_x



french_dic=pd.DataFrame(columns=["French Expression","English Expression(s)",
                                 "Occured Index(es)","Occured Groups","Tokenization with POS",
                                 "Entry with stop words",
                                 "Entry without stop words", "Stop words",
                                 "Entry (final)", 
                                 "word1", "syn_1",
                                 "word2", "syn_2",
                                 "word3", "syn_3",
                                 "Rule"])

x1=lexicalclues["Rule String FR"].value_counts()
x2=lexicalclues[["Rule String EN","Rule String FR","Grouping","Rule"]]


english_list=[]
for i in range(0,len(x1)):
    print(i)
    french_string = nlp(x1.index[i])
    
    
    french_word_pos=[]
    french_stopwords=[]
    french_word_pos_with_stopwords=[]
    french_word_pos_without_stopwords=[]
    french_word_pos_final=[]
    words=[""]*3
    words_syn=[]
    
    french_word_pos_filter_list=["verb","noun"]
    french_kept_stop_words_list=['suffisant','parler','dedans','nouveau','concernant']
    

    for num,token in enumerate(french_string):
        postag=spacy.explain(token.pos_)
        french_word_pos.append([token.text,postag])
        
        if postag in french_word_pos_filter_list:
            french_word_pos_with_stopwords.append(token.text)
            
        if (postag in french_word_pos_filter_list) and not (nlp.vocab[token.text].is_stop):
            french_word_pos_without_stopwords.append(token.text)
            french_word_pos_final.append(token.text)
            
        if (postag in french_word_pos_filter_list) and (nlp.vocab[token.text].is_stop):
            french_stopwords.append(token.text)
            if token.text in french_kept_stop_words_list:
                 french_word_pos_final.append(token.text) 
            
    
    length_with_stopwords=len(french_word_pos_with_stopwords)
    length_without_stopwords=len(french_word_pos_without_stopwords)       

    
    
    for j,value in enumerate(french_word_pos_final):
        words[j]=french_word_pos_final[j]
        words_syn.append("")

    
    for k in range(3-len(words_syn)):
        words_syn.append([])
        
    
        
    duplicated_french_words_idx=x2.index[x2["Rule String FR"] == x1.index[i]].tolist()
    duplicated_french_words_got=x2["Rule String EN"][x2["Rule String FR"] == x1.index[i]]
    duplicated_french_words_removed=list(dict.fromkeys(duplicated_french_words_got))
    
    
    duplicated_french_groupings_idx=x2.index[x2["Rule String FR"] == x1.index[i]].tolist()
    duplicated_french_groupings_got=x2["Grouping"][x2["Rule String FR"] == x1.index[i]]
    duplicated_french_groupings_removed=list(dict.fromkeys(duplicated_french_groupings_got))
    
    duplicated_french_rule_idx=x2.index[x2["Rule String FR"] == x1.index[i]].tolist()
    duplicated_french_rule_got=x2["Rule"][x2["Rule String FR"] == x1.index[i]]
    duplicated_french_rule_removed=list(dict.fromkeys(duplicated_french_rule_got))    
    
    french_dic=french_dic.append({"French Expression":x1.index[i],
                                  "Occured Index(es)":duplicated_french_words_idx ,
                                  "Occured Groups":duplicated_french_groupings_removed,
                                  "English Expression(s)":duplicated_french_words_removed,
                                  "Tokenization with POS":french_word_pos,
                                  "Entry with stop words": french_word_pos_with_stopwords,
                                  "Entry without stop words": french_word_pos_without_stopwords,
                                  "Stop words": french_stopwords,
                                  "Entry (final)": french_word_pos_final,
                                  "word1": words[0],"syn_1":[],
                                  "word2": words[1],"syn_2":[],
                                  "word3": words[2],"syn_3":[],
                                  "Rule":duplicated_french_rule_removed},
                                  ignore_index=True)
    

    
french_dic.sort_values(by=['Occured Groups'],inplace=True)
french_dic.to_excel(
    "./excels/1_Phase3_TranslationTable.xlsx")
