==> 3.2_SynonymVoting.ipynb <==
import re
import pandas as pd
import numpy as np
import ast
import operator

senses=pd.read_excel(
    "./excels/3_Phase3_Word_Sense_Voting_with_Occurances_curated_CJ2.xlsx"
    ,sheet_name="Sheet2"
)
senses["Synonyms_Voting_Rest"].fillna("[]",inplace=True)

for indx in range(len(senses["Synonyms_Voting_Rest"])):
    senses["Synonyms_Voting_Rest"][indx]=ast.literal_eval(senses["Synonyms_Voting_Rest"][indx])


senses["Synonyms_Voting"]=""
senses["Synonyms_Final"]=""
    
columns=["Syn1", "Syn2", "Syn3", "Syn4",  "Syn5", 
         "Syn6", "Syn7", "Syn8", "Syn9",  "Syn10", 
         "Syn11", "Syn12", "Syn13", "Syn14",  "Syn15", 
         "Syn16", "Syn17", "Syn18", "Syn19",  "Syn20"]

for row in range(len(senses["List_of_Occurances"])):
    list_top20=[]
    for i in range(20):
        x=senses[columns[i]][row]
        if not pd.isna(x):
            x=x.replace("'","")
            x=x.strip()
            list_top20.append(x) 
    senses["Synonyms_Voting"][row]=list_top20

senses["Synonyms_Final"]=senses["Synonyms_Voting"]+senses["Synonyms_Voting_Rest"]

for row in range(len(senses["List_of_Occurances"])):
    senses["Synonyms_Final"][row]=list(dict.fromkeys((senses["Synonyms_Final"][row])))

senses=senses[["Word","List_of_Occurances","Synonyms_Final"]]


senses.to_excel(
   "./excels/3_Phase3_Word_Sense_Voting_with_Occurances_curated_CJ2_Cleaned.xlsx")

sum([len(i) for i in senses["Synonyms_Voting_Rest"]])

sum([len(i) for i in senses["Synonyms_Voting"]])

sum([len(i) for i in senses["Synonyms_Final"]])

senses["Synonyms_Final"][0]
