==> 4_SynonymTableMerger.ipynb <==
import pandas as pd

word_table=pd.read_excel(
    "./excels/1_Phase3_TranslationTable.xlsx")

syn_table=pd.read_excel(
    "./excels/3_Phase3_Word_Sense_Voting_with_Occurances_curated_CJ2_Cleaned.xlsx")
    

word_table=word_table.fillna("")

for row_w in range(word_table.shape[0]):
    for row_s in range(syn_table.shape[0]):
        if (word_table["word1"][row_w]==syn_table["Word"][row_s]):
            word_table["syn_1"][row_w]=syn_table["Synonyms_Final"][row_s]
        if (word_table["word2"][row_w]==syn_table["Word"][row_s]):
            word_table["syn_2"][row_w]=syn_table["Synonyms_Final"][row_s]
        if (word_table["word3"][row_w]==syn_table["Word"][row_s]):
            word_table["syn_3"][row_w]=syn_table["Synonyms_Final"][row_s]
            

word_table.to_excel(
    "./excels/4_Phase3_TranslationTable.xlsx")           
    
word_table
