==> 5.1_ CFG Renaming.ipynb <==
import json


list_of_notebooks=[
"Validation_Notebook_0_100.ipynb",
"Validation_Notebook_101_200.ipynb",
"Validation_Notebook_201_300.ipynb",
"Validation_Notebook_301_400.ipynb",
"Validation_Notebook_401_500.ipynb",
"Validation_Notebook_501_600.ipynb",
"Validation_Notebook_601_737.ipynb"]


for i in range(len(list_of_notebooks)):
    with open('./notebooks-Clement/'+list_of_notebooks[i]) as json_data:
        notebook = json.load(json_data)
        json_data.close()
    
    for j in range(len(notebook["cells"])):
        if "S ->" not in notebook["cells"][j]["source"][4]:
            print (list_of_notebooks[i],j)



import json

update=[]


list_of_notebooks=[
"Validation_Notebook_0_100.ipynb",
"Validation_Notebook_101_200.ipynb",
"Validation_Notebook_201_300.ipynb",
"Validation_Notebook_301_400.ipynb",
"Validation_Notebook_401_500.ipynb",
"Validation_Notebook_501_600.ipynb",
"Validation_Notebook_601_737.ipynb"]


for i in range(len(list_of_notebooks)):
    with open('./notebooks-Clement/'+list_of_notebooks[i]) as json_data:
        notebook = json.load(json_data)
        json_data.close()
    
    for j in range(len(notebook["cells"])):
         update.append([list_of_notebooks[i],j,notebook["cells"][j]["source"][4]])


for i in range(len(update)):
     update[i][2]=update[i][2][4:].strip().split()[:-1]
update

helperList=["Article", "Patient", "Termin", "Pronoun", "Avec", "Deter", "Member", "Month", "Seasons", "Aucun", "Word", "Depuis"]

for j in range(len(update)):
    
    update[j].append([])
    indexing=1
    
    for i in range(len(update[j][2])):

        if not update[j][2][i] in helperList:
            update[j][3].append("Var"+str(indexing))
            indexing+=1
        else:
            update[j][3].append(update[j][2][i])

update        

import json


list_of_notebooks=[
"Validation_Notebook_0_100.ipynb",
"Validation_Notebook_101_200.ipynb",
"Validation_Notebook_201_300.ipynb",
"Validation_Notebook_301_400.ipynb",
"Validation_Notebook_401_500.ipynb",
"Validation_Notebook_501_600.ipynb",
"Validation_Notebook_601_737.ipynb"]

 
for i in range(len(update)):
    with open('./notebooks-Clement/'+update[i][0]) as json_data:
        working_notebook = json.load(json_data)
        json_data.close()

        src=working_notebook["cells"][update[i][1]]["source"]

        for j in range(len(src)):
            for k in range(len(update[i][2])):
                working_notebook["cells"][update[i][1]]["source"][j]=working_notebook["cells"][update[i][1]]["source"][j].replace(update[i][2][k],update[i][3][k])


    with open('./notebooks-Clement/'+update[i][0], 'w+') as outfile:
        json.dump(working_notebook, outfile)
