==> 6_ CFG2Rule.ipynb <==
import json

update=[]


list_of_notebooks=[
"Validation_Notebook_0_100.ipynb",
"Validation_Notebook_101_200.ipynb",
"Validation_Notebook_201_300.ipynb",
"Validation_Notebook_301_400.ipynb",
"Validation_Notebook_401_500.ipynb",
"Validation_Notebook_501_600.ipynb",
"Validation_Notebook_601_737.ipynb"]


for i in range(len(list_of_notebooks)):
    with open('./notebooks-Clement/'+list_of_notebooks[i]) as json_data:
        notebook = json.load(json_data)
        json_data.close()
    
    for j in range(len(notebook["cells"])):
         update.append(notebook["cells"][j]["outputs"][0]["text"])


len(update)

from functools import reduce
update=reduce(lambda x,y: x+y,update)

update

update=[x.replace(" |", "|") for x in update]
update=[x.replace(" _", "") for x in update]
# update=[x.replace(" _es", "es") for x in update]
# update=[x.replace(" _e", "e") for x in update]
# update=[x.replace(" _s", "s") for x in update]
# update=[x.replace(" _nt", "nt") for x in update]

len(update)

import xlwt
wb = xlwt.Workbook()
ws = wb.add_sheet('Rules')
ws.write(0, 0, "Trigger Rules")
for rownum, value in enumerate(update):
    ws.write(rownum+1, 0, value)
wb.save("./notebooks-Clement/List_of_Rules.xls")
