==> 7_ ResultAnalyser.ipynb <==
def get_model_result(file):
    
    import pandas as pd

    negation_df=pd.read_excel("../Fast_Context_French/conf/"+file,sheet_name='Negation')
    experiencer_df=pd.read_excel("../Fast_Context_French/conf/"+file,sheet_name='Experiencer')
    temporality_df=pd.read_excel("../Fast_Context_French/conf/"+file,sheet_name='Temporality')


    import sklearn.metrics as skm

    negation_df["gold"]=negation_df["Gold Value"]=="Negated"
    negation_df["pred"]=negation_df["FastContext Value"]!="Not negated"

    precision_neg_true=skm.precision_score(negation_df["gold"], negation_df["pred"],pos_label=True)
    recall_neg_true=skm.recall_score(negation_df["gold"], negation_df["pred"],pos_label=True)
    f1_neg_true=skm.f1_score(negation_df["gold"], negation_df["pred"],pos_label=True)
    
    
    precision_neg_false=skm.precision_score(negation_df["gold"], negation_df["pred"],pos_label=False)
    recall_neg_false=skm.recall_score(negation_df["gold"], negation_df["pred"],pos_label=False)
    f1_neg_false=skm.f1_score(negation_df["gold"], negation_df["pred"],pos_label=False)


    experiencer_df["gold"]=experiencer_df["Gold Value"]=="Patient"
    experiencer_df["pred"]=experiencer_df["FastContext Value"]=="patient"

    precision_exp_true=skm.precision_score(experiencer_df["gold"], experiencer_df["pred"],pos_label=True)
    recall_exp_true=skm.recall_score(experiencer_df["gold"], experiencer_df["pred"],pos_label=True)
    f1_exp_true=skm.f1_score(experiencer_df["gold"], experiencer_df["pred"],pos_label=True)

    precision_exp_false=skm.precision_score(experiencer_df["gold"], experiencer_df["pred"],pos_label=False)
    recall_exp_false=skm.recall_score(experiencer_df["gold"], experiencer_df["pred"],pos_label=False)
    f1_exp_false=skm.f1_score(experiencer_df["gold"], experiencer_df["pred"],pos_label=False)

    temporality_df["gold"]=temporality_df["Gold Value"]=="Historical"
    temporality_df["pred"]=temporality_df["FastContext Value"]!="Not historical"

    precision_his_true=skm.precision_score(temporality_df["gold"], temporality_df["pred"],pos_label=True)
    recall_his_true=skm.recall_score(temporality_df["gold"], temporality_df["pred"],pos_label=True)
    f1_his_true=skm.f1_score(temporality_df["gold"], temporality_df["pred"],pos_label=True)
    
    precision_his_false=skm.precision_score(temporality_df["gold"], temporality_df["pred"],pos_label=False)
    recall_his_false=skm.recall_score(temporality_df["gold"], temporality_df["pred"],pos_label=False)
    f1_his_false=skm.f1_score(temporality_df["gold"], temporality_df["pred"],pos_label=False)

    data = {'Precision':[precision_neg_true,precision_neg_false,precision_exp_true,precision_exp_false,precision_his_true,precision_his_false], 
            'Recall':[recall_neg_true,recall_neg_false,recall_exp_true,recall_exp_false,recall_his_true,recall_his_false],
            'F1':[f1_neg_true,f1_neg_false,f1_exp_true,f1_exp_false,f1_his_true,f1_his_false]
            } 
    model_outcome = pd.DataFrame(data, index=["Negated","Not Negated","Patient","Not Patient","Historical","Not Historical"]) 

    return(model_outcome)


get_model_result("HEGP_dataset_output.xlsx")

get_model_result("CepiDC_dataset_output.xlsx")

# rules_id=[146,195,197,329,330,331,332,335,336,340,341,345,378,379,380,394,407,408,411,414,421,422,423,424,425,426,462,464,506,507,527,579,581,583,637,675,677]

# rules_list=(open("../Fast_Context_French/conf/"+"context_amine.txt", "r")
#            .read()
#            .split("\n"))

# fired_rules=[rules_list[rule_id-1] for rule_id in rules_id]
# fired_rules.sort()

# outfile=open("../Fast_Context_French/conf/"+"context_mehdi.txt", "w")
# outfile.write("\n".join(fired_rules))
# outfile.close()
